DE-2794: added ids_validator.py
